+++ ts/screens/ingress/CheckBox.tsx
@@ -7,8 +7,7 @@ type Props = {
   checked: boolean;
 };
 
-const checkBoxColor = "#039BE5";
-const transparentColor = "transparent";
+const checkBoxColor = `rgba(255, 255, 255, 0.15)`;
 const checkBoxIcon = "io-tick-big";
 
 const styles = StyleSheet.create({
@@ -17,25 +16,24 @@ const styles = StyleSheet.create({
     height: 20,
     alignItems: "center",
     justifyContent: "center",
+    borderWidth: 2,
     borderColor: checkBoxColor,
-    borderWidth: 2
+    borderRadius: 4
+  },
+  checked: {
+    borderColor: "transparent",
+    backgroundColor: checkBoxColor
   }
 });
 
-export const IngressCheckBox = (props: Props) => {
-  const checkboxBackgroundColor = props.checked
-    ? checkBoxColor
-    : transparentColor;
-
-  return (
-    <View style={[styles.base, { backgroundColor: checkboxBackgroundColor }]}>
-      {props.checked && (
-        <IconFont
-          name={checkBoxIcon}
-          size={styles.base.width * 0.8}
-          color={IOColors.white}
-        />
-      )}
-    </View>
-  );
-};
+export const IngressCheckBox = (props: Props) => (
+  <View style={[styles.base, props.checked ? styles.checked : {}]}>
+    {props.checked && (
+      <IconFont
+        name={checkBoxIcon}
+        size={styles.base.width * 0.7}
+        color={IOColors.white}
+      />
+    )}
+  </View>
+);